split Presentation into Path & Row
